+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added Program.cs to run and test the plugin locally
+- Expose update frequency as settings
 
 ### Changed