+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.1.4 - 2024-01-03
+
+### Changed
+
+- Updated to LibreHardwareMonitor 0.9.3
+
 ## 0.1.3 - 2023-07-27
 
 ### Changed